pyproject.toml
[tool.poetry]
name = "tezda"
version = "0.1.0"
description = ""
authors = ["michael <michaelutoh21@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.29.6"
bcrypt = "^4.0.1"
pyjwt = "^2.8.0"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
